DL training of the CIFAR-10
DL training of the CIFAR-10 dataset with SVD support. The CIFAR-10 dataset is a particular challenge for deep learning structures and is a domain of CNNs. For this reason, training is performed using a dense network in order to describe the effect of dimension reduction. Dimension reduction via SVD (singular value decomposition) shows that network training can be performed without any significant loss of quality. Of course, this technique is not only applicable to Dense_NN.